feat: Implement functions for playing the game
Include constants that represent the number of
rounds to play, number of choices to choose from,
and choices to choose. Include global variables to
monitor player scores during the game. Implement
functions for generating CPU choice, reading user
choice, evaluating the winner of a single round,
and playing a game that evaluates the winner.