fix: Include input validation for human choice
The previous getHumanChoice() function did not have
input validation. This would result in the user
losing the current round if the given input is
invalid, without a chance to provide new input.
Add if statement to the function to evaluate if the
input is invalid, alert the user, and recall the
function to give the user another chance to provide
input.