fix: Return recursive playRound() draw result
Previously, the playRound() function's response to
a draw functioned correctly when the user
enters a new choice, but not in the
case that the user quits the game after a draw.
In the case that the user cancels the
prompt to enter a new choice after a draw, the
calling playGame() function failed to interpret
the cancelled choice. This resulted in the playGame()
function's loop to continue execution into a new
round. Include a return statement for the recursive
playRound() function call. This resolves the issue
and by returning the QUIT sentinel value when the
user cancels the game after the draw.