feat: Add playAgain() function for button response
Previously, the only way to restart the game was
to refresh the page on the browser. Add a function
that is called when the user clicks the page's
restart game button. The function should reset the
global score variables to 0, and initiate a new
game by calling the playGame() function. Now the
user can use the page's button to restart the game
after the current game ends, or after the
user quits.