refactor: Refactor the logic to utilize UI feature
Previously, the program read user text input for
the game through alert messages. Replace the input
statements with an event listener that interprets
the user's choice based on the RPS button selection.
Replace the single result DOM output with
containers to hold different output information.
Refactor the program logic to respond to user click
events on RPS buttons instead of waiting on
user text input. Now the program only initiates a
round when the user selects a button, and updates
the UI output according to the game's results.